+++ clients/crypto_loan/src/binance_sdk_crypto_loan/rest_api/rest_api.py
@@ -198,7 +198,7 @@ class CryptoLoanRestAPI:
                 Borrow Flexible Loan
 
 
-        * Only available for master account
+        * This API endpoint is available for both the master account and the sub-account.
         * You can customize LTV by entering loanAmount and collateralAmount.
 
         Weight: 6000