+++ clients/derivatives_trading_options/src/binance_sdk_derivatives_trading_options/rest_api/rest_api.py
@@ -882,7 +882,7 @@ class DerivativesTradingOptionsRestAPI:
         * This rest endpoint sets up the parameters to cancel your open orders in case of an outage or disconnection.
         * Example usage:
         Call this endpoint with a countdownTime value of 10000 (10 seconds) to turn on the auto-cancel feature. If the corresponding countdownCancelAllHeartBeat endpoint is not called within 10 seconds with the specified underlying symbol, all open orders of the specified symbol will be automatically canceled. If this endpoint is called with an countdownTime of 0, the countdown timer will be stopped.
-        * The system will check all countdowns approximately every 1000 milliseconds, **please note that sufficient redundancy should be considered when using this function**. We do not recommend setting the countdown time to be too precise or too small.
+        * The system will check all countdowns approximately every 100 milliseconds, **please note that sufficient redundancy should be considered when using this function**. We do not recommend setting the countdown time to be too precise or too small.
 
         Weight: 1