+++ clients/fiat/src/binance_sdk_fiat/rest_api/rest_api.py
@@ -155,11 +155,11 @@ class FiatRestAPI:
         """
                 Fiat Withdraw(WITHDRAW)
 
-                Submit withdraw request, in this version, we only support BRL withdrawal via bank_transfer.
+                Submit withdraw request, in this version, we support BRL,ARS,MXN withdrawal via bank_transfer.
 
         You need to call this api first, and call query order detail api in a loop to get the status of the order until this order is successful.
 
-        Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
+        Before calling this API, please ensure you have completed your KYC or KYB, activated your fiat service, and verified your destination bank account on our website.
 
         you need to bind your bank account on web/app before using the corresponding account number