+++ clients/simple_earn/src/binance_sdk_simple_earn/rest_api/rest_api.py
@@ -173,7 +173,7 @@ class SimpleEarnRestAPI:
         """
                 Get BFUSD Quota Details (USER_DATA)
 
-                Get BFUSD quota details including fast redemption quota and standard redemption quota.
+                Get BFUSD quota details including subscription quota, fast redemption quota and standard redemption quota.
 
         Weight: 150
 
@@ -527,7 +527,7 @@ class SimpleEarnRestAPI:
 
                 Get Flexible Redemption Record
 
-        *	The time between `startTime` and `endTime` cannot be longer than 3 months.
+        *	The time between `startTime` and `endTime` cannot be longer than 30 days.
         *	If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         *	If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         *	If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.
@@ -579,7 +579,7 @@ class SimpleEarnRestAPI:
 
                 Get Flexible Rewards History
 
-        *	The time between `startTime` and `endTime` cannot be longer than 3 months.
+        *	The time between `startTime` and `endTime` cannot be longer than 30 days.
         *	If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         *	If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         *	If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.
@@ -654,7 +654,7 @@ class SimpleEarnRestAPI:
 
                 Get Flexible Subscription Record
 
-        * The time between `startTime` and `endTime` cannot be longer than 3 months.
+        * The time between `startTime` and `endTime` cannot be longer than 30 days.
         * If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         * If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         * If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.
@@ -770,7 +770,7 @@ class SimpleEarnRestAPI:
 
                 Get Locked Redemption Record
 
-        * The time between `startTime` and `endTime` cannot be longer than 3 months.
+        * The time between `startTime` and `endTime` cannot be longer than 30 days.
         * If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         * If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         * If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.
@@ -821,7 +821,7 @@ class SimpleEarnRestAPI:
 
                 Get Locked Rewards History
 
-        * The time between `startTime` and `endTime` cannot be longer than 3 months.
+        * The time between `startTime` and `endTime` cannot be longer than 30 days.
         * If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         * If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         * If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.
@@ -896,7 +896,7 @@ class SimpleEarnRestAPI:
 
                 Get Locked Subscription Record
 
-        * The time between `startTime` and `endTime` cannot be longer than 3 months.
+        * The time between `startTime` and `endTime` cannot be longer than 30 days.
         * If `startTime` and `endTime` are both not sent, then the last 30 days' data will be returned.
         * If `startTime` is sent but `endTime` is not sent, the next 30 days' data beginning from `startTime` will be returned.
         * If `endTime` is sent but `startTime` is not sent, the 30 days' data before `endTime` will be returned.