+++ clients/margin_trading/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.1.0 - 2026-03-16
+
+### Changed (1)
+
+- Updated `binance-common` library to version `3.7.0`
+
 ## 6.0.0 - 2026-02-25
 
 ### Added (2)